Do not use "appdata" with plugins unless okay to use prefix of embedding app
"appdata" gets resolved to "share/appname" with KStandardDirs.
And appname will be the name of the KApplication app, whatever that is.
So with plugins one better uses "data" and explicitely writes the prefix
of the respective program (or plugin) whose data/resources should be
accessed.

REVIEW: 123392

Thanks boud for review