Remove the texture options if the texture option is unchecked on load and save
We always saved a big texture to our presets, no matter whether the
option was enabled or not. This really killed us on loading presets,
because they were all as huge as the currently selected preset!

Our presets and especially the default presets can be cleaned up now.

CCMAIL:kimageshop@kde.org