Make sure the footnote styles are found by looking them up in the style manager after loading
Also make sure we draw the footnote number (in the footnote itself) at the right baseline

Using the correct footnotstyle when creating a new footnote is a bit problemantic though, because LO doesn't
see to actually save the style (at least not when not LO default value). I've made sure we have the same
default value, bu it really is a bug in LO.

But even then we have a problem, as when the style defines a font that doesn't exist LO and Qt both find a replacement font
but unfortunately it's not the same font they choose - Not much we can do about this, and it's a more general bug anyway.

BUG: 323232