Dynamic sensors: don't create translated strings uselessly
A dynamic sensor would create three translated strings (minimum label,
maximum label, id's name) on creation. We create a bunch of sensors on
stroke start, and all those translations caused a visible lag starting
a stroke on Windows.

This patch replaces the KoID with an enum except for the gui, and only
gets translated labels in the gui. Note that changing e.g. time length
doesn't update the curve max label yet unless you enable/disable the
sensor -- but that isn't a regression, that never worked.