Fix for crash when opening an image or a template
My previous fix made impossible to open images without crashing.
This fix moves the adding of the document to the KisPart, and so
the launching of the documentOpened event, after the image
is set to the document, since updateWindowMenu, called due to that
event, needs it to be not null.