BUG:344610 Don't reset the number of favorite presets to 10
The problem was that the old way of saving favorite presets (by
name) saved a config entry called favoritePresets. We also stored
the number of slots in a config entry called favoritePresets. If
we'd try to restore that number, chances are we'd be reading a
string with preset names instead.

It's kinda hard to interpret "Fuzzy_textured" as a number... So we'd
go back to the default of 10.