Fix issue with import PDF dialog resolution
The resolution selected in the import dialog wasn't set in the generated KisImage,
ending up with an image of the right size but with wrong dpi.

Removed, from the import dialog, the possibility to change horizontal and vertical
resolution separately, since we don't give this possibility anywhere else in the
program and since we always show it as a single value.

CCBUG: 348507