Brush editor Stamp and Clipboard refactoring
The Stamp and Clipboard tabs are removed and they become buttons
in the Predefined tab, so that the user must add a predefined
brush tip to use it first.
This solves some issues and inconsistencies with
temporary brush tips by removing the need of them.

Fixes a crash that happened when selecting a temporary brush
after adding the respective clone to the predefined or
when deleting the cloned brush.

Fixes a crash when attempting to add a Clipboard brush tip when the
clipboard is empty.

BUG: 345195
REVIEW: 124048
CCMAIL: kimageshop@kde