Fixed a bug when the tools became disabled after new document creation
The bug happened because all our infinity managers were processing all
the events. It happens because they were connected to the global events
queue and did not check the destination of the event. Now they check the
destination and they cannot eat any event of the different canvas.

BUG:348533
Fixes T378