FEATURE: Show the tool options in a popup
This can be set in the general settings tab: either we show
the tool options in a docker or in a popup that can be shown
using a button in the toolbar. The popup is detachable.

We need a nice icon for the tool button!

The tool option panels pops up when clicked with by pressing
(by default) the backslash key.

CCMAIL:kimageshop@kde.org