Finish continuation of the transformation action
Now the user can edit/compare transformations incrementally:

1) Transform your piece of image
2) Apply transformation
3) Click on the canvas with the transfomration tool again

Now the previously applied transformation is *continued*

If you cancel/reset your transformation, it will be reset to the
correct state, that is to the partial transformation where it has
been started.

If you switch the transformation mode during a continued action,
the previous action will be baked into the image and a new-mode-stroke
will be started.

BUG:331708




CC:kimageshop@kde.org