Fixed Undo/Redo when working with animated frames
This patch just makes undo just not crash and work correctly
when the frameId does not equal to the frameId of the operation
performed.

In the end of the command an explicit frame switch signal is emitted
to switch to the changed frame. This behavior will be defined deprecated
when we make all the frame switches undoable.