BUG:345619 All Drag & Drop of masks
The default bounds associated with a mask need to know about the
image. When we load a dropped mask, we create it with a temporary
image to reconstruct the hierarchy, and then we need to "internalize"
the nodes to the target image.

That was only done for layers, because we said

// Only layers know about the image

that was a lie...