Fix a severe crash in Transformation Masks
We should be really careful with makeCloneFrom() calls. There is a
simple rule of thumb:

"Use makeCloneFrom or makeCloneFromRough if and only of you are the
only owner of the paint device and you are 100% sure no other thread
has access to it"

The reason is simple. These call change offset and a color space of the
paint device, so all the other users with get a SIGSEGV with a non-zero
probability. If you want to copy data to a device, which is shared with
someone else, use KisPainter::copyAreaOptimized() instead. It is totlly
safe.

BUG:349819