Port away from using KIO in our filters
Done in the 2.9 branch to keep the code as close as possible to what
we have in frameworks; in frameworks, kio asserts if there is no sycoca
which means we have to drop using kio.

We lose network transparent file access, but a) I am sure that no artist
was using that anyway and b) that only works with the KDE file dialog
which is available only to a minority of users and c) having that code
separately in every filter was a mistake anywhere, we weren't consistent
so some filters had the code and others didn't.