Make cs-convert UI attempt to automatically determine when to uncheck optimise
This will have the optimisation flag turned off when you open the window from
a linear TRC colorspace. This should reduce the amount of user-error when
converting from a linear TRC to any other.

I couldn't get it to work for the other way around, but I am also not sure
if that's really necessary...