Add some more build defines
Global defines:

  -DQT_USE_QSTRINGBUILDER
  -DQT_STRICT_ITERATORS
  -DQT_NO_SIGNALS_SLOTS_KEYWORDS

For libs and krita:

  -DQT_NO_URL_CAST_FROM_STRING

I cannot manage to fix all the places where a string is cast to
a QUrl in all of Calligra, though I fixed one egregious oopsie
in braindump.

I also would like the QString and QByteArray ascii cast defines,
but that's a huge battle!