Add skeleton class for porting from KStandardDirs
KoResourcePaths has the relevant bits of API from KStandardDirs
but will use QStandardPaths internally. Only the API we actually
use in Calligra is copied.

The class is easy to use: it's a singleton and you use the static
methods.

For now, KoResourcePaths uses KStandardDirs internally, that's
on my todo for tomorrow...