Remove rng2cpp and writeodf
This is a good idea, but superceded by later experiments, was never
fully implemented, causes complications when building and is only
used in parts of the odf stack that aren't important for Krita.

See http://conference.libreoffice.org/assets/Conference/Aarhus/Slides/JosVanDenOeverblasien.pdf
for what I surmise to be writeodf's successor.