Use device->peek inside canRead methods of kimgio kra and ora plugins
KZip were used inside canRead(device) method of ora and kra handlers.
Those methods are called by the plugin capabilities(...) method. The
documentation of QImageIOPlugin::capabilities specifies that the
device state should not be changed inside this method.

The problem is that KZip (KArchive) was altering and closing the device
at destruction, causing the most visible effect of missing icons on kde
applications and a lot of "QIODevice::seek (QBuffer): The device is
not open" messages on the terminal.

Now both handler canRead methods use device->peek to ensure that the
device state is not changed when used inside the plugin capabilities
method.