Make the raw plugin build again
Since libkcdraw is obviously a digikam-only thing, not a shared library meant
to be shared among applications, we cannot depend on its api and have to keep
a local fork of it. This is one 3rd-party library a distribution cannot un-
bundle.

This is the state of libkcdraw as it was the moment before the settings widget
was removed with the message

"
 no need i18n to test code
 code cleanup
"

BUG:354048