Let changing the onion skins sliders by Shift+dragging over the widget
Just press Shift and drag over the sliders and they'll be ordered accordingly.
If you also press Ctrl or right mouse button, the sliders will move smoothly.

This feature doesn't work for the top buttons, but if you want to disable
some skins, just drag over the bottom part of the widget and when the sliders
reach 0 level, the buttons will be disabled automatically

Fixes:T1043