Add XCB tablet event handler.
Summary:
This replaces the old XInput handling code with Qt5 compatible calls.
It is a very early snapshot, and breaks everything. For some reason, I
cannot use my mouse wheel when it is enabled. Also, when I move the
stylus too fast, it gets wacky.

If the idea is worth pursuing, I think I should push this to a testing branch, since I may not be able to finish it myself.

Reviewers: rempt, dkazakov

Differential Revision: https://phabricator.kde.org/D542