Port away from `new QAction` retaining i18n strings
Summary:
This replaces most calls to `new QAction` with `KisActionRegistry::makeQAction()`, getting rid of static data in the code.

Now, instead of i18n() calls in the code right next to each action, krita.action  (and the other .action files) become the primary source of strings.

To handle this, we make a small modification to the script extractrc, which I've called action_i18n.pl.  This is intended for testing, the ultimate goal being to merge the changes to upstream extractrc.

Messages.sh is modified to include action_i18n.pl for the generation of translation strings.

Depends on D478

Tasks: T947

Reviewers: kossebau, rempt

Differential Revision: https://phabricator.kde.org/D479