Use the icon's cacheKey to map between icon and name
We need to know the name of an icon to switch between light and dark,
but iconName is only filled if the icon is loaded from a theme. Which
we don't do, because we load the icons from a qrc file. And there's
no setIconName, for whatever reason. So, use the cachekey to keep track
of the mapping.