Implemented a "Drop Frames" mode for Krita and made it default option
Now you can switch on the "Drop Frames" mode in the Animation Docker
to ensure your animation is playing with the requested frame rate,
even when the GPU cannot handle this amount of data to be shown.

TODO: we still need an icon for this feature! Atm it uses "Auto Frame" icon!

BUG:356245
CC:kimageshop@kde.org