Fix selection of the node items when doing multiple Ctrl+D&D
Qt has a bug: when we Ctrl+click on an item, the item's selections gets
toggled on mouse *press*, whereas usually it is done on mouse release.
Therefore the user cannot do a Ctrl+D&D with the default configuration.
This patch fixes the value returned by selectionCommand() callback to
fix this issue.