BUG:357987 Don't crash on loading the given file
If the clone layers clone a layer above the clones in the stack,
the original isn't available during loading. Krita shouldn't recalculate
the project at this point, but merging already starts before the
original is available, so add a temporary dummy fallback paint
device that can be used until the original becomes available.