Fix kimageformats compiling on Windows
kimageformats project needs to import some symbols from the
openexr/ilmbase, the OPENEXR_DLL define is needed.

Also updated the readme to suggest which CMake version to use (currently
all version beyond 3.3.2 seems to have bugs on Windows, they give internal
errors while configuring), to be clearer that commands needs to be run
inside a console with properly set environment variables for the compiler
and to give an alternative to MSBuild/NMake, which is jom, to have faster
builds.