Removed the perspective grid tool. Use Perspective assistant instead.
There is still an undecided part. Should we keep the Perspective
Correction feature of the Clone Brush? It is the only thing that
depended from the old tool, and I'm not sure if it was used by someone.

Porting of it might take up to 4 hours of work, should I do it or I
should better spend this time on Grids snapping?

CC:kimageshop@kde.org