Fix setting the progress updater on the import/export manager
There is one import/export manager per document, and it's a permanent
object. It used to be created with the document's progress updater,
which would be 0 at that point. Then we'd create a new progress updater
on loading; when that happens we didn't set the progress updater on
the KisImportExportManager (because that was done in the constructor
only), which meant that the import/export filters never would show
any progress.

Next: creater and use a progress updater when saving.