Fix crash with popups and tablet input handling on Windows
When having two levels of popup/modal dialogs and tapping with the
pen outside the last one so that it closes, the input is still
"redirected" to the just closed/destroyed one.
The crash can be tested opening the favorite preset popup and
then clicking to the bottom right menu icon so that the second popup
shows up, then clicking back to a favorite preset.
The simple fix is to use QPointer so we know when a QWidget has been
destroyed.