Greater-alpha blend mode
This adds a blend mode that uses a sigmoid blending function
to take the greater alpha of either the current layer or the brush.
The result of this is that making multiple passes across the
same area but with different strokes will not further increase
the opacity past the brush setting or stylus pressure. This
can help with consistently shading a region across multiple
separate strokes without creating spurious or accidental
sharp edges due to edge-overlaps with existing brush strokes.

The patch was created by Nicholas Guttenberg

Differential revision: https://phabricator.kde.org/D666