Add gamma as hotkey-steps configuration to the config.
This'll allow you to change the gamma of the HSY advanced color selector,
to turn it linear for example.

The hotkey config will allow setting the amount of steps that affect the L and K
and the similar redder/greener/yellower/bluer/hue/sat actions. These are still
hardcoded to use a gamma of 2.2 for now.

Ref T210