Correct some wrong initialization sizes and leaks
The source pixel buffer should be as big as the pixelSize of the source color space.

channelValuesF it's used for the source color space and the destination color space,
so it has to be of a size equal to the bigger of the two color spaces.
Given that the biggest one will have at most (and only) 5 channels,
we keep it fixed at 5.

Differential Revision: https://phabricator.kde.org/D1136