Create KisMimeDatabase: a wrapper around QMimeDatabase
This wrapper has a fallback for mimetypes that Qt doesn't know about,
and which we cannot reliably add to the share/mime on Windows and
OSX. For now, these patterns are hard-coded in KisMimeDatabase, but
they should come from the import/export plugins. The problem is that
they then should be defined in the .json files, and those aren't
translated yet.

Closes D1176