Remove all usage of QUrl from the import/export filters
The import/export filters always work on local files, so there
is no need to convert the inputFile() to a QUrl, only to convert
it back to a local filename again. This fixes a number of issues
where we handled the conversion between QUrl and a filename the
wrong way, and it also simplifies our code a lot.