Fix colorspace registry leaking
Multiple threads could find that the colorspace wasn't cached and decide
that they had to create it, so when taking the write lock we check if
the colorspace isn't actually already in the cache.
This check fixes also a second issue where, not having the colorspace
factory cache, and since the profileName might be different after having
checked the cache, we might create a colorspace that was already in the cache.

A practical example is the colorspace profile "XYZ identity built-in", which
we say it's in the cache, then a colorspace using "XYZ identity built-in (Default)"
is searched with the colorSpace function. That profile name happens to be an alias
of the one in the cache, but we first search for the aliased name and not finding it,
without this patch, we could decide that we have to create a new colorspace and
overwrite the other one in the cache, doing this everytime that colorspace is
searched.

The patch also optimize the access to the caches using an iterator so that
we know if a colorspace is present and eventually take its pointer without
multiple accesses to the cache.