krita_xgettext.sh: Fix modification of multiline msgctxt tags
Summary:
For a message like this:
  msgctxt ""
  "context string"
  msgid "original text"
the former code changed the msgctxt and also added
  msgctxt "(qtundo-format)"
because it did not find the existing msgctxt on the previously line
w.r.t. msgid.

To fix the problem, we additionally check:
  lastLine !~ /^"/

Thanks to Luigi Toscano for noticing and for his patch that does
basically the same.

CCMAIL: luigi.toscano@tiscali.it

Reviewers: rempt, ltoscano

Reviewed By: rempt, ltoscano

Differential Revision: https://phabricator.kde.org/D1203