Fix infinite loops when using Clones+TransformMasks+Animation+LoD
KisRegenerateFrameStrokeStrategy runs in the LoD mode (splits
into Lod0+LodN parts), therefore, Lod0 part is wrapped into
Suspend/ResumeUpdates strokes, which collect all the updates it
initiates and issues them after all the Lod0 strokes are finished.

The problem is that KisRegenerateFrameStrokeStrategy doesn't want this
events collecting! It has already blocked the GUI updates and waits until
the projection is ready to be passed to the animation cache. No filtering
or postponing should happen.

I bet you are asking how Clone Layers are related? Well, the point is that
in normal situation the projection update filters installed by the suspend/
resume strokes are not used at all when regenerating the frame. The
regeneration stroke just doesn't call any KisImage update methods. Instead,
it creates KisAsyncMerger directly, so the filters are bypassed. But...

When KisAsyncMerger finds a node with clones, it initiates updates all the
clones as well, by simply calling clone->setDirty(), which effectively
ends up in the updates filter.

Now we just explicitly disable all the filters in the frame regeneration
stroke.

Fixes T1208