Find all the languages so the language chooser is shown
Make sure that on all platforms the local xdg_data_dirs is part
of the environment variable, the one that is bin'../share or
MacOS/../share.

And remove the kauthorized stuff as well. Nobody is going to disable
anything in the help menu in krita in any kind of kiosk situation.