Update krita to compile against Vc >= 1.0.0
This updates krita to work with Vc version 1.0.0 and above. Some changes to the
build system were needed as the compile options are now set differently. The
compile flags are set on the lib which uses the functionality. This is needed in
a non packager build only. Compile the KisCompositionBenchmark against the
current available Vc optimization.

Thanks to Vir for all his help.

Tested packager and non packager build. Run the and compared the output of
./KisCompositionBenchmark.shell. The speed is nearly the same. For avx a
different code path is used when avx2 is not available which makes it tick
slower but correct in all cases according to Vir.

Compiled and tested against vc 1.0.0 and vc 1.2.0.

(cherry picked from commit b68c1c7666510e268babc4cd87cd9d5ad34ff2d5)