Remove QUrl from KisImportExportManager
QUrl, the spiritual descendant of KUrl, was used here because someone
had wanted to make the import/export manager network transparent. That
was never finished, just like KoStore never was network transparent:
it's KParts (which we don't use) which is network transparent.

So all we were doing here was String->QUrl->QString all the time.