Make all plugins modules
Otherwise they don't load in an mingw build. This means that the
unittests that _link_ to a plugin don't build on OSX and MSVC anymore,
but that's a small price to pay, so make those tests conditional.

Of course we should spend time making these plugins into a plugin +
shared library, but I don't have time for that now.