Fix painting onion skins in LoD mode
This patch adds a special cache object that renders all the
onion skins on a separate device and activated LoD on it.

That cache not only fixes ghost strokes appearing on the canvas,
but also reduces the amount of work that have to be done when
rendering the onion skins. So it the end, painting speed with
onion skins should become higher.

Fixes T1141
BUG:356086