Fix a "bended lines"+AMD+Qt5.6 bug
The bug is actually in Qt 5.6. They decided to implement a motion event
compression without *any* way to disable it. So we have a lot of events
dropped resulting in "bended lines" on canvas.

For some reason the bug is reproducible on AMD GPUs only, with openGL
activated. It seems like either Qt asks for some synchronization, or
rendering on AMD is just too slow. I don't know the exact reason,
but being busy with rendering makes Qt drop a lot of events.

Since this event compression cannot be disabled in any official way, we
should start patching Qt on Linux as well. Yes, it means that Krita 3.0
cannot be distributed through official distribution repositories. Which
means that Krita Lime will have to be shut down as soon as Ubuntu 16.04
will upgrade to Qt 5.6.

CC:kimageshop@kde.org
CC:Shawn.Rutledge@theqtcompany.com
BUG:361619
Fixes T2224