Fix masks update problem
1) Create a paint layer (layer1) filled with color
2) Create a paint layer (layer2) with anything
3) Add a transparency mask to layer1 (tmask1)
4) Make tmask1 invisible

Layer1's projection is now filled with non-cropped color

5) Make layer1 invisible
6) Make layer2 invisible

7) Make tmask1 visible
Layer1's projection is still filled with non-cropped
color, because it is invisible and its updates are locked.

8) Make layer1 visible
The *cropped* area is rendered on screen, layer1'a
projection is still wrong!

9) Make layer2 visible
Wrong layer1's projection is now painted on screen -> BUG

This patch fixes the problem by re-rendering the
layer1's projection at step 8.

BUG:349484,345612
Ref T2085