T2311 fix theme switching on Windows
Remove the extra default theme that was linked to empty data.
Ignore KDE theming on windows as it doesn't work right.
On startup, add an indicator to the theme that is being used

I am not sure if the KDE default theme is working on other OSs, so that might need to be ignored for those as well.